+++ android/src/main/java/com/getcapacitor/community/database/sqlite/SQLite/BiometricListener.java
@@ -1,0 +1,8 @@
+package com.getcapacitor.community.database.sqlite.SQLite;
+
+import androidx.biometric.BiometricPrompt;
+
+public interface BiometricListener {
+    void onSuccess(BiometricPrompt.AuthenticationResult result);
+    void onFailed();
+}