+++ ios/Plugin/SqliteConfig.swift
@@ -7,4 +7,7 @@
 
 public struct SqliteConfig {
     var iosDatabaseLocation: String?
+    var biometricAuth: Int?
+    var biometricTitle: String?
+    var iosKeychainPrefix: String?
 }